+++ packages/http/src/mocker/negotiator/InternalHelpers.ts
@@ -19,7 +19,10 @@ export function findExampleByKey(httpContent: IHttpContent, exampleKey: string) 
 }
 
 export function hasContents(v: IHttpOperationResponse): v is PickRequired<IHttpOperationResponse, 'contents'> {
-  return !!v.contents;
+  // TODO: adjust after https://github.com/stoplightio/http-spec/pull/25 is merged
+  const x = v.contents || [];
+
+  return !!x.filter(y => y.mediaType !== '*/*').length;
 }
 
 export function findBestHttpContentByMediaType(