+++ packages/http/src/mocker/negotiator/__tests__/NegotiatorHelpers.spec.ts
@@ -16,6 +16,17 @@ import { IHttpNegotiationResult, NegotiationOptions } from '../types';
 const chance = new Chance();
 const logger = createLogger('TEST', { enabled: false });
 
+const assertPayloadlessResponse = (actualResponse: Either.Either<Error, IHttpNegotiationResult>) => {
+  assertRight(actualResponse, response => {
+    expect(response).not.toHaveProperty('bodyExample');
+    expect(response).not.toHaveProperty('mediaType');
+    expect(response).not.toHaveProperty('schema');
+
+    expect(response).toHaveProperty('code');
+    expect(response).toHaveProperty('headers');
+  });
+};
+
 function anHttpOperation(givenHttpOperation?: IHttpOperation) {
   const httpOperation = givenHttpOperation || {
     method: chance.string(),
@@ -84,6 +95,24 @@ describe('NegotiatorHelpers', () => {
       });
 
       describe('and has no static contents', () => {
+        it('returns an empty payload response for an invalid request', () => {
+          httpOperation = anHttpOperation(httpOperation)
+            .withResponses([
+              {
+                code: actualCode,
+                headers: [],
+                contents: [],
+              },
+            ])
+            .instance();
+
+          const actualResponse = helpers.negotiateOptionsForInvalidRequest(httpOperation.responses, ['422', '400'])(
+            logger,
+          );
+
+          assertPayloadlessResponse(actualResponse);
+        });
+
         test('and has schemable contents should return first such contents', () => {
           httpOperation = anHttpOperation(httpOperation)
             .withResponses([
@@ -146,9 +175,7 @@ describe('NegotiatorHelpers', () => {
             logger,
           );
 
-          assertLeft(negotiationResult, e =>
-            expect(e.message).toBe('Neither schema nor example defined for 422 response.'),
-          );
+          assertRight(negotiationResult, e => expect(e).toStrictEqual({ code: '422', headers: [] }));
         });
       });
     });
@@ -527,12 +554,26 @@ describe('NegotiatorHelpers', () => {
         });
       });
 
-      describe('httpContent does not exist ', () => {
+      describe('the response exists, but there is no httpContent', () => {
         const httpResponseSchema: IHttpOperationResponse = {
-          code: chance.integer({ min: 100, max: 599 }).toString(),
+          code: '200',
           contents: [],
-          headers: [],
         };
+
+        it('returns an empty payload response', () => {
+          const desiredOptions: NegotiationOptions = {
+            dynamic: false,
+            mediaTypes: ['*/*'],
+          };
+
+          const actualResponse = helpers.negotiateOptionsBySpecificResponse(
+            { ...httpOperation, responses: [{ code: '200', contents: [{ mediaType: 'text/plain ' }] }] },
+            desiredOptions,
+            httpResponseSchema,
+          )(logger);
+
+          assertPayloadlessResponse(actualResponse);
+        });
 
         it('should throw an error', () => {
           const desiredOptions: NegotiationOptions = {
@@ -549,21 +590,33 @@ describe('NegotiatorHelpers', () => {
 
           expect(Either.isLeft(actualResponse)).toBeTruthy();
         });
+      });
 
-        it('should return a payload-less response', () => {
+      describe('the content-type from Accept header cannot be matched', () => {
+        it('returns 406', () => {
+          const desiredOptions: NegotiationOptions = {
+            dynamic: false,
+            mediaTypes: ['application/json'],
+          };
+
+          const httpResponseSchema: IHttpOperationResponse = {
+            code: '200',
+            contents: [
+              {
+                mediaType: 'text/plain',
+              },
+            ],
+          };
+
           const actualResponse = helpers.negotiateOptionsBySpecificResponse(
-            { ...httpOperation, method: 'head' },
-            { dynamic: false, mediaTypes: ['*/*'] },
-            { code: '200' },
+            { ...httpOperation, responses: [{ code: '200', contents: [{ mediaType: 'text/plain ' }] }] },
+            desiredOptions,
+            httpResponseSchema,
           )(logger);
 
-          expect(Either.isRight(actualResponse)).toBeTruthy();
-
-          assertRight(actualResponse, response => {
-            expect(response).not.toHaveProperty('bodyExample');
-            expect(response).not.toHaveProperty('mediaType');
-            expect(response).not.toHaveProperty('schema');
-          });
+          assertLeft(actualResponse, e =>
+            expect(e.message).toBe('The server cannot produce a representation for your accept header'),
+          );
         });
       });
     });